requirements.txt
jupyter==1.0.0
newspaper3k==0.2.8
nltk==3.6.5
pdfservices-sdk==1.0.1
sumy==0.9.0
transformers==4.11.3